Additional HTTP status codes.
